refactor(api-v3): Move quest routes to separate file
